Moved files around to match config, separated model builders and transformers #396
